Leer CSV de un centro desde URL

La funcion importa a Rstudio el dataset del centro que queramos desde github, SIN DESCARGAR EL ARCHIVO

Usage

read_centro_csv(centro)

Arguments

centro

(str) Nombre del centro a descargar

Value

Si la descarga fue exitosa devuelve el dataset

See also

Author

JM y VM

Examples

NH0910 <- read_centro_csv("NH0910")
#> Rows: 15553 Columns: 35
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr   (1): id
#> dbl   (5): temperatura_abrigo_150cm, temperatura_abrigo_150cm_maxima, temper...
#> lgl  (28): temperatura_intemperie_5cm_minima, temperatura_intemperie_50cm_mi...
#> date  (1): fecha
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
NH0910
#> # A tibble: 15,553 × 35
#>    id     fecha      temperatura_abrigo_150cm temperatura_abrigo_150cm_maxima
#>    <chr>  <date>                        <dbl>                           <dbl>
#>  1 NH0910 1975-01-01                     21                              31  
#>  2 NH0910 1975-01-02                     21.2                            28.5
#>  3 NH0910 1975-01-03                     24.2                            32.5
#>  4 NH0910 1975-01-04                     23.5                            31  
#>  5 NH0910 1975-01-05                     19                              22  
#>  6 NH0910 1975-01-06                     22.2                            27.5
#>  7 NH0910 1975-01-07                     24.2                            30  
#>  8 NH0910 1975-01-08                     26.5                            33.5
#>  9 NH0910 1975-01-09                     26.5                            35  
#> 10 NH0910 1975-01-10                     19.2                            25  
#> # ℹ 15,543 more rows
#> # ℹ 31 more variables: temperatura_abrigo_150cm_minima <dbl>,
#> #   temperatura_intemperie_5cm_minima <lgl>,
#> #   temperatura_intemperie_50cm_minima <lgl>,
#> #   temperatura_suelo_5cm_media <lgl>, temperatura_suelo_10cm_media <lgl>,
#> #   temperatura_inte_5cm <lgl>, temperatura_intemperie_150cm_minima <lgl>,
#> #   humedad_suelo <lgl>, precipitacion_pluviometrica <dbl>, granizo <lgl>, …